requirements.txt
# チラシ商品価格抽出システム v1.0 - 依存関係
# Python 3.10+ required

# 画像処理ライブラリ
opencv-python>=4.8.0
Pillow>=10.0.0

# OCR APIクライアント
google-cloud-vision>=3.4.0

# データ処理ライブラリ
pandas>=2.0.0
numpy>=1.24.0

# AI/LLM APIクライアント
openai>=1.0.0

# テストライブラリ
pytest>=7.4.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0

# 設定管理
python-dotenv>=1.0.0

# 代替API依存関係（オプション）
# azure-cognitiveservices-vision-computervision>=0.9.0  # Azure Computer Vision API
# anthropic>=0.18.0  # Anthropic Claude API

# 開発用ツール（オプション）
# black>=23.0.0  # コードフォーマッター
# flake8>=6.0.0  # リンター
# mypy>=1.0.0    # 型チェッカー